refactor: move issue #71 regression test into functional_test.go
Delete standalone bug71_test.go and consolidate the test into the
existing functional test suite alongside other integration tests.
